requirements.txt
gradio==3.24.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
scikit_learn==1.2.2